refactor(ops): extract project index domain from coding-agent
Refs #933
