+++ crates/tau-ops/src/lib.rs
@@ -1,7 +1,9 @@
 mod channel_store_admin;
 mod daemon_runtime;
+mod project_index;
 mod transport_health;
 
 pub use channel_store_admin::*;
 pub use daemon_runtime::*;
+pub use project_index::*;
 pub use transport_health::*;
